fixed setting float values on moderation

+++ app/models/target.rb
@@ -13,7 +13,7 @@ class Target
   money_field :fixed_prices_bronze
 
   field :prc_price, type: Integer, default: 0
-  field :prc_prices_bronze, type: Integer, default: 0
+  field :prc_prices_bronze, type: Float, default: 0.0
 
 	symbolize :confirm_mode, :in => [:auto, :manual], :default => :auto
 	field :confirm_url, type: String
